Changed Long to Integer customTime

+++ src/main/java/ua/chekmaryov/barber_stat/dto/barberOffering/BarberOfferingDtoResponse.java
@@ -12,6 +12,6 @@ public record BarberOfferingDtoResponse(
         Long offerId,
         String offerName,
         BigDecimal price,
-        Long customTime
+        Integer customTime
 ) {
 }
